create server package from main.go
Enable other packages to use slack handlers.
