+++ server/resp.go
@@ -1,0 +1,24 @@
+package server
+
+import (
+	"encoding/json"
+	"log"
+	"net/http"
+)
+
+func jsonResp(w http.ResponseWriter, msg string) {
+	w.Header().Set("Content-Type", "application/json; charset=utf-8")
+	resp := map[string]string{"text": msg}
+	r, err := json.Marshal(resp)
+	if err != nil {
+		log.Println("Couldn't marshal hook response:", err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	w.Write(r)
+}
+
+func plainResp(w http.ResponseWriter, msg string) {
+	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Write([]byte(msg))
+}
